requirements.txt
fastapi==0.64.0
celery==4.4.7
redis==3.5.3
pydantic==1.8.2
SQLAlchemy==1.4.18
plotly==4.14.3
uvicorn==0.13.0
psycopg2-binary==2.9.1
pandas==1.2.4
requests==2.25.0